workiing on splitting codec
